Introduce persistent counting of tickets over device's lifetime.
A JSON file is used to hold the state of the device over restarts. Ticket number can be printed on the ticket and ist stored in evidence.
A per device ID can be set to distinguish tickets and evidence from individual VMSE2000 Units. Device ID will be stored in evidence and can be printed on tickets.
Also fix a crash that could happen if a missing token file was tried to be opened.